feat(ssrRequest): 处理SSE连接/markdown显示/语法高亮等

+++ src/request/sseRequest.js
@@ -1,9 +1,10 @@
 import { fetchEventSource } from "@microsoft/fetch-event-source";
-import { getAccessToken } from '@/utils/tools';
+import { getAccessToken, getRefreshToken, setAccessToken, setRefreshToken } from '@/utils/tools';
 import { message } from 'antd';
+import { getNewToken } from "./auth.js";
 
+function sseRequest(url, params, onMessage, setIsLoading, onMyError) {
 
-function sseRequest(url, params, onMessage) {
   let eventSource = null;
   const ctrl = new AbortController();
 
@@ -29,42 +30,52 @@ function sseRequest(url, params, onMessage) {
       async onopen(response) {
         if (response.ok && response.status === 200) {
           console.log('连接成功')
+          // 此处response 里面返回的code可能等于500
           return
         } else if (response.status === 401) {
+          console.log('token过期')
 
 
-          console.log('登录身份过期')
-          message.error('身份认证过期，请重新登录');
+          let res = await getNewToken(getRefreshToken())
+          if (res.code === 200) {
+            // token刷新成功
+            const newAccessToken = res.data.access_token;
+            const newRefreshToken = res.data.refresh_token;
+            // 本地更新token
+            setAccessToken(newAccessToken);
+            setRefreshToken(newRefreshToken);
 
-
-
-          setTimeout(() => {
-            window.location.replace('/auth');
-            sessionStorage.clear();
-          }, 1000);
-
-
+            // setIsLoading(true) //重新等待
+            // startSse(); //重新请求
+            setTimeout(() => {
+              setIsLoading(true); // 重新设置等待状态
+              startSse(); //重新请求
+            }, 0);
+          } else {
+            console.log('token更新失败')
+            message.error('身份认证过期，请重新登录');
+            setTimeout(() => {
+              window.location.replace('/auth');
+              sessionStorage.clear();
+            }, 1000);
+          }
 
         } else {
           console.log('连接异常')
           message.error(response.error || '连接异常');
+          onMyError(response) //手动抛出一个错误
         }
       },
-      async onmessage(event) {
+      onmessage(event) {
+        // console.log("🚀 ~ onmessage ~ event:", event)
         const data = JSON.parse(event.data);
-        console.log("🚀 ~ onmessage ~ data:", data)
-        // if (onMessage) {
-        //   onMessage(data);
-        // }
         if (Array.isArray(data)) {
           data.forEach(messageData => {
-            // 处理单条消息
             if (onMessage) {
               onMessage(messageData);
             }
           });
         } else {
-          // 处理单条消息
           if (onMessage) {
             onMessage(data);
           }
@@ -73,18 +84,22 @@ function sseRequest(url, params, onMessage) {
 
       },
 
-      async onerror(error) {
+      onerror(error) {
         // 关闭 SSE 连接
-        eventSource.close();
+        // eventSource.close();
         console.log(error || '连接错误');
         message.error(error || '连接错误');
+        throw error
       },
-      async onclose() {
+      onclose() {
+        setIsLoading(false);
         console.log('关闭连接');
       },
     });
   }
-  startSse();
+  startSse()
+
+
   return eventSource;
 }
